separate pruner interface and config
