+++ internal/cache/gomap/gomap.go
@@ -11,14 +11,13 @@ import (
 )
 
 type mapCache[I, K comparable, V any] struct {
-	config         *cache.Config
-	indexMap       map[I]*index[K, V]
-	closeAutoPrune func()
+	config   *cache.CacheConfig
+	indexMap map[I]*index[K, V]
 }
 
 // NewCache returns an in-memory Cache implementation based on the builtin go map type.
 // Object values are stored as-is and there is no encoding or decoding involved.
-func NewCache[I, K comparable, V any](background context.Context, indices []I, config cache.Config) cache.Cache[I, K, V] {
+func NewCache[I, K comparable, V any](background context.Context, indices []I, config cache.CacheConfig) cache.Cache[I, K, V] {
 	m := &mapCache[I, K, V]{
 		config:   &config,
 		indexMap: make(map[I]*index[K, V], len(indices)),
@@ -29,7 +28,6 @@ func NewCache[I, K comparable, V any](background context.Context, indices []I, c
 			entries: make(map[K]*entry[V]),
 		}
 	}
-	m.closeAutoPrune = config.StartAutoPrune(background, m)
 	return m
 }
 
@@ -91,13 +89,12 @@ func (c *mapCache[I, K, V]) Clear(_ context.Context) error {
 }
 
 func (c *mapCache[I, K, V]) Close(ctx context.Context) error {
-	c.closeAutoPrune()
 	return ctx.Err()
 }
 
 type index[K comparable, V any] struct {
 	mutex   sync.RWMutex
-	config  *cache.Config
+	config  *cache.CacheConfig
 	entries map[K]*entry[V]
 }
 
@@ -158,7 +155,7 @@ type entry[V any] struct {
 	lastUse atomic.Int64 // UnixMicro time
 }
 
-func (e *entry[V]) isValid(c *cache.Config) bool {
+func (e *entry[V]) isValid(c *cache.CacheConfig) bool {
 	if e.invalid.Load() {
 		return false
 	}
