+++ internal/cache/gomap/gomap_test.go
@@ -43,7 +43,7 @@ func (o *testObject) Value() *testObject {
 }
 
 func Test(t *testing.T) {
-	c := NewCache[testIndex, string, *testObject](context.Background(), testIndices, cache.Config{
+	c := NewCache[testIndex, string, *testObject](context.Background(), testIndices, cache.CacheConfig{
 		Name:       "tester",
 		MaxAge:     time.Second,
 		LastUseAge: time.Second / 4,
